feat(market): enhance market item handling for staff purchases
