+++ game/Code/UI/Menus/TabMenu/Sections/MarketTabMenuSection.razor.scss
@@ -41,6 +41,7 @@ MarketTabMenuSection {
     padding: $space-2 $space-4;
     background-color: $color-primary;
     border-radius: $space-3;
+    border: 1px solid transparent;
     cursor: pointer;
     font-size: $text-sm;
     color: $color-gray-400;
@@ -54,6 +55,12 @@ MarketTabMenuSection {
     &.active {
       color: $color-accent;
       background-color: rgba($color-accent, 0.08);
+    }
+
+    &.admin-toggle.active {
+      color: #e04c4c;
+      background-color: rgba(#e04c4c, 0.08);
+      border-color: #e04c4c;
     }
   }
 
@@ -130,6 +137,7 @@ MarketTabMenuSection {
 
   // Card
   .entity-card {
+    position: relative;
     display: flex;
     flex-direction: row;
     align-items: stretch;
@@ -155,22 +163,17 @@ MarketTabMenuSection {
         box-shadow: none;
       }
     }
+  }
 
-    &.is-admin-spawn {
-      .card-inner {
-        background-color: rgba($color-accent, 0.1);
-        outline: 1px solid rgba($color-accent, 0.28);
-      }
-
-      .admin-lock i {
-        color: $color-accent;
-        opacity: 0.7;
-      }
-
-      .card-price {
-        color: $color-green-400;
-      }
-    }
+  .card-staff-border {
+    position: absolute;
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    border: 1px solid #e04c4c;
+    border-radius: $space-2;
+    pointer-events: none;
   }
 
   // Entity color as a subtle tinted layer
